pyproject.toml
[project]
name = "subscriptionchurnrisk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "datasets>=4.8.4",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
]
